# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MPI API Documentação",
        "description": "Documentação da API do sistema MPI",
        "contact": {
            "email": "luizleao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/processo-sei": {
            "get": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Listar todos os processos SEI",
                "operationId": "5ab17202f3143a034977874da3683bd5",
                "responses": {
                    "200": {
                        "description": "Lista de processos SEI",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProcessoSei"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Criar um novo processosei",
                "operationId": "4f72c712a2e9df6ecf603b6d0699e64d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idTipoProcessoSei",
                                    "idConflito",
                                    "numero",
                                    "assunto",
                                    "especificacao",
                                    "interessado"
                                ],
                                "properties": {
                                    "idTipoProcessoSei": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "idConflito": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123.456.789.123"
                                    },
                                    "assunto": {
                                        "type": "string",
                                        "example": "Descrição do assunto"
                                    },
                                    "especificacao": {
                                        "type": "string",
                                        "example": "Detalhar especificação"
                                    },
                                    "interessado": {
                                        "type": "string",
                                        "example": "João da Silva, Maria José,..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Processo SEI criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessoSei"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/processo-sei/{id}": {
            "get": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Obter um processo SEI específico",
                "operationId": "84df39349d4a4b78acda88662609c22d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do processo sei"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Atualizar um processo SEI específico",
                "operationId": "03ebc92a61a9fa12075a352e1c663dd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idTipoProcessoSei",
                                    "idConflito",
                                    "numero",
                                    "assunto",
                                    "especificacao",
                                    "interessado"
                                ],
                                "properties": {
                                    "idTipoProcessoSei": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "idConflito": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123.456.789.123"
                                    },
                                    "assunto": {
                                        "type": "string",
                                        "example": "Descrição do assunto"
                                    },
                                    "especificacao": {
                                        "type": "string",
                                        "example": "Detalhar especificação"
                                    },
                                    "interessado": {
                                        "type": "string",
                                        "example": "João da Silva, Maria José,..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Processo SEI atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Excluir um processo Sei específico",
                "operationId": "171ccb1866145cbbc271ac909242ff16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Processo SEI excluído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/processo-sei/conflito/{idConflito}": {
            "get": {
                "tags": [
                    "ProcessosSei"
                ],
                "summary": "Listar todos os processos SEI de um Conflito",
                "operationId": "c34a22efadd55b25d876ccca94010e2b",
                "parameters": [
                    {
                        "name": "idConflito",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de processos SEI"
                    },
                    "400": {
                        "description": "Termo de pesquisa não fornecido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProcessoSei": {
                "properties": {
                    "idProcessoSei": {
                        "type": "integer",
                        "example": "1"
                    },
                    "tipoProcessoSei": {
                        "$ref": "#/components/schemas/TipoProcessoSei"
                    },
                    "numero": {
                        "type": "string",
                        "example": "59004.000960/2024-68"
                    },
                    "assunto": {
                        "type": "string",
                        "example": "Conflito"
                    },
                    "especificacao": {
                        "type": "string",
                        "example": "Especificação do processo"
                    },
                    "interessado": {
                        "type": "string",
                        "example": "Luiz Silva"
                    }
                },
                "type": "object"
            },
...
            "TipoProcessoSei": {
                "properties": {
                    "idTipoProcessoSei": {
                        "type": "integer",
                        "example": "2"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Tipo de Processo B"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}